requirements.txt
# Core data handling and analysis
numpy
pandas
scipy

# API interaction
requests
python-dotenv

# Plotting (optional, but often useful)
matplotlib
seaborn

# Trading specific (examples, can be changed later)
# ccxt  # For interacting with cryptocurrency exchanges
# alpaca-trade-api # For interacting with Alpaca

# Utilities
python-dateutil
pytz

# Testing
pytest
pytest-cov

# Linting and formatting (optional, good practice)
# flake8
# black
# isort

# Technical Analysis
ta

# Generative AI for Sentiment Analysis
google-generativeai
